+++ packages/blocks/core/CHANGELOG.md
@@ -105,6 +105,7 @@
 - `Page List` block: support improved.
 - `Post Author` block: support improved.
 - `Post Author Name` block: support improved.
+- `Post Content` block: support improved.
 
 
 ### Automated Tests
@@ -147,6 +148,7 @@
 - Updated E2E tests to check support of `Page List` block.
 - Updated E2E tests to check support of `Post Author` block.
 - Updated E2E tests to check support of `Post Author Name` block.
+- Updated E2E tests to check support of `Post Content` block.
 
 
 ## 1.3.0 (2025-04-27)